Move codelens utils
