+++ extension/src/extension.ts
@@ -1,7 +1,7 @@
 /* The extension */
 import { EmulatorState, EmulatorController } from './emulator/emulator-controller'
 import { CommandController } from './commands/command-controller'
-import { refreshCodeLenses } from './utils/utils'
+import { refreshCodeLenses } from './utils/codelens'
 import { Account } from './emulator/account'
 import { UIController } from './ui/ui-controller'
 import { ExtensionContext } from 'vscode'
