--- extension/src/utils/strings.ts
@@ -1,6 +1,0 @@
-// Preffix to indicate which account is active
-export const ACTIVE_PREFIX = '🟢'
-export const INACTIVE_PREFIX = '⚫️'
-export const ADD_NEW_PREFIX = '🐣'
-
-export const CREATE_NEW_ACCOUNT = 'Create New Account'
